Collapse openai instrumentations to single project
