+++ instrumentation/openai-client-instrumentation/instrumentation-1.0/src/main/java/co/elastic/otel/openai/v0_14/OpenAiOkHttpClientBuilderInstrumentation.java
