+++ instrumentation/openai-client-instrumentation/instrumentation-1.0/src/main/java/co/elastic/otel/openai/wrappers/InstrumentedEmbeddingsService.java
