+++ instrumentation/openai-client-instrumentation/instrumentation-1.0/src/test/java/co/elastic/otel/openai/ResponseHeaderScrubber.java
