+++ instrumentation/openai-client-instrumentation/instrumentation-1.0/src/test/resources/mappings/chattest.yaml
