+++ instrumentation/openai-client-instrumentation/instrumentation-1.0/src/test/resources/mappings/embeddingstest.yaml
