+++ settings.gradle.kts
@@ -17,10 +17,7 @@ include("agentextension")
 include("bootstrap")
 include("custom")
 include("instrumentation")
-include("instrumentation:openai-client-instrumentation:common")
-include("instrumentation:openai-client-instrumentation:testing-common")
-include("instrumentation:openai-client-instrumentation:instrumentation-0.8")
-include("instrumentation:openai-client-instrumentation:instrumentation-0.14")
+include("instrumentation:openai-client-instrumentation:instrumentation-1.0")
 include("inferred-spans")
 include("resources")
 include("runtime-attach")
